Remove identifier sync back in v-repeat
The auto-sync back when an identifier is used causes issues
such as #213 and #234, while the feature is a rarely used
edge case that is also a bit magical. It can be easily circumvented
by using object arrays instead of primitive ones, and the sync-back
is better left in userland for those who really need it.

For reference, Angular also doesn't support sync back for primitive values.